Clean up deprecated files, swap 24x24 to 28x28 train/test images
