--- src/not_for_release/dsen2/models/README.md
@@ -1,5 +1,0 @@
-## Models
-
-The DSen2 models are included in the repository.
-
-Due to size restrictions, the VDSen2 (very deep) models can be downloaded from [s2_033_lr_1e-04.hdf5](http://n.ethz.ch/~lanarasc/DSen2/s2_033_lr_1e-04.hdf5) and [s2_034_lr_1e-04.hdf5](http://n.ethz.ch/~lanarasc/DSen2/s2_034_lr_1e-04.hdf5).
