--- src/not_for_release/dsen2/testing/README.md
@@ -1,6 +1,0 @@
-## Testing
-
-Find a demo in `demoDSen2.py`.
-
-
-Use the file `s2_tiles_supres.py` to super-resolve Sentinel-2 tiles, by giving the `.zip` file or the `.xml` file of the unzipped file. You must also provide an output filename, the default is geotiff (extension `.tif`).
